See discussion in https://github.com/pytorch/pytorch/pull/50320#discussion_r554447365.