@peterjc123 2021-01-17
Timings:
Without this PR:
CUDA10.1_TEST1: 32m 36s
CUDA10.1_TEST2: 2h 2m 27s

With this PR:
CUDA10.1_TEST1: 1h 19m 49s
CUDA10.1_TEST2: 1h 19m 5s